# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Exam Website API Documentation",
        "description": "Documentation done using Swagger",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://3.143.249.185/api",
            "description": "Smart Exam API Server"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in as Student or as Instructor",
                "description": "login to your account",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "LoggedIn successfully"
                                        },
                                        "auth": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Email Or Password is incorrect"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log out From your Account",
                "description": "logout from your account",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "Successfully Logged Out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You Have Successfully Logged Out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/changePassword": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "changing password",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/forgotPassword": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resetting password",
                "operationId": "resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password has been successfully changed!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request to reset password",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reset link sent!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/verifyEmail": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request to verify email",
                "operationId": "verifyEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyEmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account verified!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ChangePasswordRequest": {
                "title": "changePasswordRequest",
                "description": "change password request body data",
                "required": [
                    "currentPassword",
                    "newPassword"
                ],
                "properties": {
                    "currentPassword": {
                        "title": "currentPassword",
                        "description": "User Current Password",
                        "type": "string",
                        "example": "12345678"
                    },
                    "newPassword": {
                        "title": "newPassword",
                        "description": "User New Password",
                        "type": "string",
                        "example": "Aaa45678"
                    }
                },
                "type": "object"
            },
            "ForgotPasswordRequest": {
                "title": "ForgotPasswordRequest",
                "description": "Forgot password request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "example": "example@example.com"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "firstName": {
                        "title": "First Name",
                        "description": "First name of User",
                        "type": "string",
                        "example": "Laila"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of User",
                        "type": "string",
                        "example": "h@ex.com"
                    },
                    "lastName": {
                        "title": "Last Name",
                        "description": "Last name of User",
                        "type": "string",
                        "example": "Mohsen"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of User",
                        "type": "string",
                        "example": "123456"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "gender of User",
                        "type": "string",
                        "example": "Male"
                    },
                    "image": {
                        "title": "image",
                        "description": "image of User",
                        "type": "string",
                        "example": "http://pixels.com/4040"
                    },
                    "type": {
                        "title": "type",
                        "description": "type of User",
                        "type": "string",
                        "example": "Admin"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of User",
                        "type": "string",
                        "example": "012233344455"
                    },
                    "verified": {
                        "title": "verified",
                        "description": "verified User",
                        "type": "boolean",
                        "example": "true"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "email_verified_at": {
                        "title": "Email verified at",
                        "description": "Email verified at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "ResetPasswordRequest": {
                "title": "ResetPasswordRequest",
                "description": "Reset password request body data",
                "required": [
                    "email",
                    "token",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "token": {
                        "title": "token",
                        "description": "Reset password url token",
                        "type": "string",
                        "example": "64 char string"
                    },
                    "password": {
                        "title": "password",
                        "description": "new password",
                        "type": "string",
                        "example": "yuiuiui"
                    }
                },
                "type": "object"
            },
...
            "StoreLoginRequest": {
                "title": "StoreLoginRequest",
                "description": "Store Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of the user",
                        "type": "string",
                        "example": "omar@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the user",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
...
            "VerifyEmailRequest": {
                "title": "VerifyEmailRequest",
                "description": "Email verification request body data",
                "required": [
                    "email",
                    "code"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "code": {
                        "title": "code",
                        "description": "Vericication code",
                        "type": "string",
                        "example": "6C3BT5"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}